Support sorting based on values from a dictionary.
Sorting is implemented by storing the parent dictionary's ColKey and the specific dictionary key we want to sort by.

This is a work-in-progress, the performance characteristics of searching the dictionary for the desired key on every iteration of the sort predicate are unknown.

There dictionary ColKey and the desired sort key are currently stored inside the ColKey structure to avoid changing too much of the SortDescriptor interface.

Sorting by dictionary keys using keyPath is currently not implemented (but desired).

In the future this could be extended to support sorting on specific indexes of a collection, ie. sort('array[0]').